Brain Segmentation/requirements.txt
flask
torch
torchvision
pillow
numpy
opencv-python
werkzeug